Extracts base functionality to super class for panels
