+++ src/Leaderboard.js
@@ -1,4 +1,4 @@
-import Panel from './Panel';
+import Panel from './ui/Panel';
 import { Graphics, Text } from 'pixi.js'
 import { delayed_call } from "./utils";
 
@@ -69,8 +69,8 @@ export default class Leaderboard extends Panel {
     if (width < 200) {
       width = 200;
     }
-    if (this._box.width !== width || this._box.height !== height) {
-      this.resize(width, height);
+    if (this.resize(width, height)) {
+      this.draw();
       this._title.x = 0.5 * (this.width - this._title.width);
     }
   }
