+++ src/ui/Control.js
@@ -1,8 +1,8 @@
 import { Graphics, Rectangle } from 'pixi.js';
 
-export default class Panel extends Graphics {
+export default class Control extends Graphics {
   _onResize = null;
-  _box = new Rectangle;
+  _box = new Rectangle();
 
   constructor(view, config) {
     super();
@@ -20,17 +20,13 @@ export default class Panel extends Graphics {
 
   resize(width, height) {
     if (this._box.width === width && this._box.height === height) {
-      return;
+      return false;
     }
     this._box.width = width;
     this._box.height = height;
-    this.clear();
-    this.lineStyle(this._config.lineStyle);
-    this.beginFill(this._config.fill[0], this._config.fill[1]);
-    this.drawRect(0, 0, width, height);
-    this.endFill();
     if (this._onResize) {
       this._onResize();
     }
+    return true;
   }
 }
